Restructure NavBar, Signup Page, Header, fix CSS for login/signup
