+++ client/assets/_login.scss
@@ -3,14 +3,9 @@
  * http://bit.ly/1ywfhq7
  */
 
- .login-page,
- .signup-page {
-   height: 100vh;
-
-   .login-form,
-   .signup-form {
-     margin-top: 100px;
-   }
+ .login-form,
+ .signup-form {
+   margin-top: 100px;
 
    form {
      margin-top: 10px;
