+++ client/components/sessions/SessionContainer.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 
-import Header from '../Header';
+import Header from '../molecules/Header';
 import * as reviewActions from '../../actions/reviewActions';
 import * as itemActions from '../../actions/itemActions';
